llamatry/pyproject.toml
[tool.poetry]
name = "llamatry"
version = "0.1.0"
description = ""
authors = ["Jason <jason@jxnl.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.6"
opentelemetry-api = "^1.17.0"
opentelemetry-sdk = "^1.17.0"
opentelemetry-instrumentation = "^0.38b0"
urllib3 = "<2.0"
opentelemetry-instrumentation-requests = "^0.38b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
